Fix crash in ControllerMgr and cast to uintptr in VertexArray
* ControllerMgr::initializeDefault() may have caused a crash because it
reads a file then passes it to glfwUpdateGamepadMappings, BUT that will
read until it sees a null terminator, so if there isn't one then it will
go out of bounds. Because the last character happens to be a line break,
I decided to change the last character to a null so that it would stop
when it's supposed to.

* Cast to uintptr_t in VertexArray::process()